move functions from step.js to enemy.js
